+++ packages/backend/server/src/crypto/abstract-key-provider.ts
@@ -54,27 +54,13 @@ export abstract class AbstractKeyProvider {
         // Create the key id
         const keyId = `kcc-signing-${Date.now()}-${webcrypto.randomUUID()}`;
 
-        const extraProps = {
-            use: 'sig',
-            alg: alg,
-            kid: `kcc-signing-${Date.now()}-${webcrypto.randomUUID()}`,
-        }
-
-        const publicJwk = {
-            ...extraProps,
-            ...await jose.exportJWK(keyPair.publicKey),
-        };
-
-        const privateJwk = {
-            ...extraProps,
-            ...await jose.exportJWK(keyPair.privateKey),
-        };
-
         // Build a connector keys object
         return {
             kid: keyId,
-            privateJwk: privateJwk,
-            publicJwk: publicJwk,
+            alg: alg,
+            use: 'sig',
+            publicKey: keyPair.publicKey,
+            privateKey: keyPair.privateKey,
         };
     }